Removed useGameBoardActions.js
